refactor(protocol): Moved buffer handling out of protocol.
