+++ lua/smuggler/protocol.lua
@@ -6,36 +6,6 @@ local config = require("smuggler.config")
 local snitch = require("smuggler.snitch")
 local buffers = require("smuggler.buffers")
 
-function M.socketsdir()
-	if vim.fn.has("unix") or vim.fn.has("mac") then
-		return "/run/user/" .. tostring(uv.getuid()) .. "/julia/replsmuggler/"
-	elseif vim.fn.has("win32") then
-		return "\\\\.\\pipe\\"
-	else
-		error("Unsupported platform.")
-	end
-end
-
-function M.getavailablesockets()
-	local directory = M.socketsdir()
-	local res = {}
-	for v in vim.fs.dir(directory) do
-		res[#res + 1] = directory .. v
-	end
-	return res
-end
-
-function M.choosesocket()
-	local sockets = M.getavailablesockets()
-	local choice = nil
-	vim.ui.select(sockets, {
-		prompt = "Select a socket:",
-	}, function(c)
-		choice = c
-	end)
-	return choice
-end
-
 function M.serialize_requests(bufnbr)
 	config.debug("Started serializer.")
 	local bufconfig = config.buf[bufnbr]
@@ -152,40 +122,6 @@ function M.runclient(bufnbr)
 		bufconfig.session_connected_event.wait()
 		M.treat_incoming(bufnbr)
 	end)
-end
-
-function M.bufconfig(bufnbr, force, settings)
-	if bufnbr == nil then
-		bufnbr = vim.api.nvim_get_current_buf()
-	end
-	if force == nil then
-		force = false
-	end
-	if settings == nil then
-		settings = { evalbyblocks = config.eval_by_blocks }
-	end
-
-	local current_config = config.buf[bufnbr]
-
-	if current_config ~= nil then
-		local closed = current_config.socket:is_closing()
-		if not closed and not force then
-			return 0
-		elseif not closed and force then
-			current_config.socket:close()
-		end
-	end
-
-	local socket_path = M.choosesocket()
-	if socket_path == nil then
-		return -1
-	end
-	config.buf[bufnbr] = buffers.buffer(bufnbr, socket_path, socket, settings)
-
-	nio.run(function()
-		M.runclient(vim.api.nvim_get_current_buf())
-	end)
-	return 0
 end
 
 function M.send(code, firstline, filename)
